2-lightweight-charts/requirements.txt
numpy
python-binance
TA-Lib
backtrader
flask